FIX: unequal-length lists work when exploding
polyfill(resolution, explode=True) did not work when
the resulting lists had different length (which would
typically be the case)

I think this may be due to API change in Pandas?
It would be strange since this breaks the notebook
01_unified_data_layers.ipynb

Either way, it is now fixed + a test that would catch
it was added.